dynamic crypto providers
