+++ apps/web/app/global.css
@@ -2,6 +2,9 @@
 @tailwind components;
 @tailwind utilities;
 
+@import '@rainbow-me/rainbowkit/styles.css';
+@import '@coinbase/onchainkit/styles.css';
+
 /* For Webkit-based browsers (Chrome, Safari and Opera) */
 :not(.scrollbar)::-webkit-scrollbar {
   display: none;
